day6 + c++23 madness

+++ template.cpp
@@ -6,18 +6,17 @@
 #include <array>
 #include <string_view>
 #include <ranges>
-#include <print>
 #include <algorithm>
 #include <format>
 #include <optional> // monads!
 #include <functional>
 
 
-int part1(std::filesystem::path file){
+int64_t part1(std::filesystem::path file){
     return 0;
 }
 
-int part2(std::filesystem::path file){
+int64_t part2(std::filesystem::path file){
     return 0;
 }
